nav restructure
 - initialization handling
 - namespace is now defined in the route
