+++ src/components/DataTimeline/TimelineItemWrapper.tsx
@@ -15,7 +15,7 @@
 // limitations under the License.
 
 import React from 'react';
-import { ITimelineItem } from '../../interfaces';
+import { ITimelineItem } from '../../_core/interfaces';
 import { OppositeTimelineItem } from './TimelineItems/OppositeTimelineItem';
 import { TimelineItem } from './TimelineItems/TimelineItem';
