+++ src/components/DataTimeline/TimelineItems/OppositeTimelineItem.tsx
@@ -18,41 +18,40 @@ import {
   TimelineConnector,
   TimelineContent,
   TimelineDot,
-  TimelineItem as TItem,
+  TimelineItem,
   TimelineOppositeContent,
   TimelineSeparator,
 } from '@mui/lab';
 import makeStyles from '@mui/styles/makeStyles';
 import React from 'react';
-import { ITimelineItem } from '../../../interfaces';
+import { ITimelineItem } from '../../../_core/interfaces';
 import { TimelineContentPanel } from './TimelineContentPanel';
 
 interface Props {
   item: ITimelineItem;
 }
 
-export const TimelineItem: React.FC<Props> = ({ item }) => {
+export const OppositeTimelineItem: React.FC<Props> = ({ item }) => {
   const classes = useStyles();
 
   return (
     <>
-      <TItem>
-        <TimelineOppositeContent>{item.time}</TimelineOppositeContent>
+      <TimelineItem>
+        <TimelineOppositeContent>
+          <TimelineContentPanel
+            title={item.title}
+            description={item.description}
+            onClick={item.onClick}
+          />
+        </TimelineOppositeContent>
         <TimelineSeparator>
           <TimelineDot className={classes.dot}>
             {item.icon ? item.icon : undefined}
           </TimelineDot>
           <TimelineConnector />
         </TimelineSeparator>
-        <TimelineContent>
-          <TimelineContentPanel
-            title={item.title}
-            description={item.description}
-            onClick={item.onClick}
-            color="secondary"
-          />
-        </TimelineContent>
-      </TItem>
+        <TimelineContent>{item.time}</TimelineContent>
+      </TimelineItem>
     </>
   );
 };
