--- src/components/interfaces.tsx
@@ -1,4 +1,0 @@
-export interface IDataWithHeader {
-  header: string;
-  data: string | number;
-}
