+++ src/interfaces/navigation.ts
@@ -1,0 +1,14 @@
+import { RouteObject } from 'react-router-dom';
+
+export interface FFRouteObject extends RouteObject {
+  icon?: JSX.Element;
+  name?: string;
+  path: string;
+  children?: FFRouteObject[];
+}
+
+export const NAMESPACES_PATH = 'namespaces';
+export const ACTIVITY_PATH = 'activity';
+export const EVENTS_PATH = 'events';
+export const TRANSACTIONS_PATH = 'transactions';
+export const OPERATIONS_PATH = 'operations';
